feat: optimize perps WebSocket rendering with atom updates and component refactoring
- Optimize atom updates to prevent unnecessary re-renders when value unchanged
- Add computed atom for MMR calculation to reduce redundant computations
- Refactor PerpTickerBar into separate desktop/mobile components for better maintainability
- Extract PerpTokenSelectorMobile component for mobile-specific token selection
- Add proper type safety for IPerpsActiveAssetCtxAtom updates
- Add demo atoms for Jotai testing and validation
