+++ packages/kit-bg/src/states/jotai/utils/wrapAtomPro.ts
@@ -67,6 +67,9 @@ export function wrapAtomPro(
             prev: any | Promise<any>,
           ) => any | Promise<any> | typeof JOTAI_RESET
         )(prevValue);
+        if (prevValue === nextValue) {
+          return;
+        }
       }
 
       if (nextValue instanceof Promise) {
