+++ packages/kit/src/views/Perp/components/TickerBar/PerpTickerBar.tsx
@@ -1,560 +1,17 @@
-import { memo, useCallback } from 'react';
+import { memo } from 'react';
 
-import { useIntl } from 'react-intl';
+import { useMedia } from '@onekeyhq/components';
 
-import {
-  Badge,
-  DebugRenderTracker,
-  Divider,
-  Icon,
-  IconButton,
-  NumberSizeableText,
-  Popover,
-  ScrollView,
-  SizableText,
-  SkeletonContainer,
-  Tooltip,
-  XStack,
-  YStack,
-  useMedia,
-} from '@onekeyhq/components';
-import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
-import {
-  usePerpsActiveAccountSummaryAtom,
-  usePerpsActiveAssetAtom,
-  usePerpsActiveAssetCtxAtom,
-} from '@onekeyhq/kit-bg/src/states/jotai/atoms';
-import { ETranslations } from '@onekeyhq/shared/src/locale';
-import { EModalRoutes } from '@onekeyhq/shared/src/routes';
-import { EModalPerpRoutes } from '@onekeyhq/shared/src/routes/perp';
-import {
-  NUMBER_FORMATTER,
-  formatDisplayNumber,
-} from '@onekeyhq/shared/src/utils/numberUtils';
-
-import { useFundingCountdown, usePerpSession } from '../../hooks';
-import { useMMR } from '../../hooks/useMMR';
-import { PerpSettingsButton } from '../PerpSettingsButton';
-import { PerpTokenSelector } from '../TokenSelector/PerpTokenSelector';
-import { PerpsAccountNumberValue } from '../TradingPanel/components/PerpsAccountNumberValue';
-
-function useTickerBarIsLoading() {
-  const { isReady, hasError } = usePerpSession();
-  const [assetCtx] = usePerpsActiveAssetCtxAtom();
-  const { markPrice } = assetCtx?.ctx || {
-    markPrice: '0',
-  };
-  return !isReady || hasError || parseFloat(markPrice) === 0;
-}
-
-function TickerBarMarkPrice() {
-  const intl = useIntl();
-  const [assetCtx] = usePerpsActiveAssetCtxAtom();
-  const formattedMarkPrice = assetCtx?.ctx?.markPrice || '';
-  const isLoading = useTickerBarIsLoading();
-  return (
-    <DebugRenderTracker
-      name="TickerBarMarkPrice"
-      position="bottom-left"
-      offsetY={10}
-    >
-      <SkeletonContainer isLoading={isLoading} width={80} height={28}>
-        <Tooltip
-          placement="top"
-          renderTrigger={
-            <SizableText size="$headingXl">{formattedMarkPrice}</SizableText>
-          }
-          renderContent={
-            <SizableText size="$bodySm">
-              {intl.formatMessage({
-                id: ETranslations.perp_mark_price_tooltip,
-              })}
-            </SizableText>
-          }
-        />
-      </SkeletonContainer>
-    </DebugRenderTracker>
-  );
-}
-
-function TickerBarChange24hPercent() {
-  const intl = useIntl();
-  const { gtMd } = useMedia();
-  const [assetCtx] = usePerpsActiveAssetCtxAtom();
-  const change24hPercent = assetCtx?.ctx?.change24hPercent || 0;
-  const isLoading = useTickerBarIsLoading();
-
-  return (
-    <DebugRenderTracker
-      name="TickerBarChange24hPercent"
-      position="bottom-right"
-      offsetY={10}
-    >
-      <SkeletonContainer isLoading={isLoading} width={50} height={16}>
-        <NumberSizeableText
-          size={gtMd ? '$headingXs' : '$bodySmMedium'}
-          fontSize={gtMd ? undefined : 10}
-          mt={gtMd ? undefined : '$-2'}
-          color={change24hPercent >= 0 ? '$green11' : '$red11'}
-          formatter="priceChange"
-          formatterOptions={{
-            showPlusMinusSigns: true,
-          }}
-        >
-          {change24hPercent}
-        </NumberSizeableText>
-      </SkeletonContainer>
-    </DebugRenderTracker>
-  );
-}
-
-function TickerBarOraclePrice() {
-  const intl = useIntl();
-  const [assetCtx] = usePerpsActiveAssetCtxAtom();
-  const formattedOraclePrice = assetCtx?.ctx?.oraclePrice || '';
-  const isLoading = useTickerBarIsLoading();
-
-  return (
-    <DebugRenderTracker name="TickerBarOraclePrice">
-      <YStack>
-        <Tooltip
-          renderTrigger={
-            <SizableText
-              size="$bodySm"
-              borderBottomWidth="$px"
-              borderTopWidth={0}
-              borderLeftWidth={0}
-              borderRightWidth={0}
-              borderBottomColor="$border"
-              borderStyle="dashed"
-              color="$textSubdued"
-            >
-              {intl.formatMessage({
-                id: ETranslations.perp_token_bar_oracle_price,
-              })}
-            </SizableText>
-          }
-          renderContent={
-            <SizableText size="$bodySm">
-              {intl.formatMessage({
-                id: ETranslations.perp_oracle_price_tooltip,
-              })}
-            </SizableText>
-          }
-          placement="top"
-        />
-        <SkeletonContainer isLoading={isLoading} width={80} height={16}>
-          <SizableText size="$headingXs">{formattedOraclePrice}</SizableText>
-        </SkeletonContainer>
-      </YStack>
-    </DebugRenderTracker>
-  );
-}
-
-function TickerBar24hVolume() {
-  const intl = useIntl();
-  const [assetCtx] = usePerpsActiveAssetCtxAtom();
-  const volume24h = assetCtx?.ctx?.volume24h || '0';
-  const isLoading = useTickerBarIsLoading();
-
-  return (
-    <DebugRenderTracker name="TickerBar24hVolume">
-      <YStack>
-        <SizableText size="$bodySm" color="$textSubdued">
-          {intl.formatMessage({
-            id: ETranslations.perp_token_bar_24h_Volume,
-          })}
-        </SizableText>
-        <SkeletonContainer isLoading={isLoading} width={80} height={16}>
-          <SizableText size="$headingXs">
-            $
-            {formatDisplayNumber(
-              NUMBER_FORMATTER.marketCap(volume24h.toString()),
-            )}
-          </SizableText>
-        </SkeletonContainer>
-      </YStack>
-    </DebugRenderTracker>
-  );
-}
-
-function TickerBarOpenInterest() {
-  const intl = useIntl();
-  const [assetCtx] = usePerpsActiveAssetCtxAtom();
-  const { openInterest = '0', markPrice = '0' } = assetCtx?.ctx || {};
-  const isLoading = useTickerBarIsLoading();
-
-  return (
-    <DebugRenderTracker name="TickerBarOpenInterest">
-      <YStack>
-        <Tooltip
-          renderTrigger={
-            <SizableText
-              size="$bodySm"
-              color="$textSubdued"
-              borderBottomWidth="$px"
-              borderTopWidth={0}
-              borderLeftWidth={0}
-              borderRightWidth={0}
-              borderBottomColor="$border"
-              borderStyle="dashed"
-            >
-              {intl.formatMessage({
-                id: ETranslations.perp_token_bar_open_Interest,
-              })}
-            </SizableText>
-          }
-          renderContent={
-            <SizableText size="$bodySm">
-              {intl.formatMessage({
-                id: ETranslations.perp_open_interest_tooltip,
-              })}
-            </SizableText>
-          }
-          placement="top"
-        />
-        <SkeletonContainer isLoading={isLoading} width={80} height={16}>
-          <SizableText size="$headingXs">
-            $
-            {formatDisplayNumber(
-              NUMBER_FORMATTER.marketCap(
-                (Number(openInterest) * Number(markPrice)).toString(),
-              ),
-            )}
-          </SizableText>
-        </SkeletonContainer>
-      </YStack>
-    </DebugRenderTracker>
-  );
-}
-
-function TickerBarFundingRate() {
-  const intl = useIntl();
-  const countdown = useFundingCountdown();
-  const [assetCtx] = usePerpsActiveAssetCtxAtom();
-  const fundingRate = assetCtx?.ctx?.fundingRate || '0';
-  const isLoading = useTickerBarIsLoading();
-  return (
-    <DebugRenderTracker name="TickerBarFundingRate">
-      <YStack>
-        <Tooltip
-          renderTrigger={
-            <SizableText
-              size="$bodySm"
-              borderBottomWidth="$px"
-              borderTopWidth={0}
-              borderLeftWidth={0}
-              borderRightWidth={0}
-              borderBottomColor="$border"
-              borderStyle="dashed"
-              color="$textSubdued"
-            >
-              {intl.formatMessage({
-                id: ETranslations.perp_token_bar_Funding,
-              })}
-            </SizableText>
-          }
-          renderContent={
-            <YStack gap="$2">
-              <SizableText size="$bodySm">
-                {intl.formatMessage({
-                  id: ETranslations.perp_funding_rate_tip1,
-                })}
-              </SizableText>
-              <SizableText size="$bodySm">
-                {intl.formatMessage({
-                  id: ETranslations.perp_funding_rate_tip2,
-                })}
-              </SizableText>
-            </YStack>
-          }
-          placement="top"
-        />
-        <SkeletonContainer isLoading={isLoading} width={120} height={16}>
-          <XStack alignItems="center" gap="$2">
-            <Tooltip
-              renderTrigger={
-                <XStack alignItems="center" gap="$2">
-                  <SizableText
-                    size="$headingXs"
-                    color={parseFloat(fundingRate) >= 0 ? '$green11' : '$red11'}
-                  >
-                    {(parseFloat(fundingRate) * 100).toFixed(4)}%
-                  </SizableText>
-                  <SizableText size="$headingXs" color="$text">
-                    {countdown}
-                  </SizableText>
-                </XStack>
-              }
-              renderContent={
-                <YStack gap="$1">
-                  <YStack py="$1" gap="$0.5" justifyContent="space-between">
-                    <SizableText size="$bodySm" color="$textSubdued">
-                      {intl.formatMessage({
-                        id: ETranslations.perp_ticker_annualized_funding_tooltip,
-                      })}
-                    </SizableText>
-                    <SizableText
-                      size="$bodySmMedium"
-                      color={
-                        parseFloat(fundingRate) >= 0 ? '$green11' : '$red11'
-                      }
-                    >
-                      {(parseFloat(fundingRate) * 100 * 24 * 365).toFixed(2)}%
-                    </SizableText>
-                  </YStack>
-                  <Divider />
-                  <YStack py="$1" gap="$0.5" justifyContent="space-between">
-                    <SizableText size="$bodySm" color="$textSubdued">
-                      {intl.formatMessage({
-                        id: ETranslations.perp_trades_history_direction,
-                      })}
-                    </SizableText>
-                    {parseFloat(fundingRate) >= 0 ? (
-                      <SizableText size="$bodySmMedium" color="$text">
-                        <SizableText size="$bodySmMedium" color="$green11">
-                          {intl.formatMessage({
-                            id: ETranslations.perp_ticker_direction_funding_tooltip_long,
-                          })}
-                        </SizableText>{' '}
-                        {intl.formatMessage({
-                          id: ETranslations.perp_ticker_direction_funding_tooltip_pays,
-                        })}{' '}
-                        <SizableText size="$bodySmMedium" color="$red11">
-                          {intl.formatMessage({
-                            id: ETranslations.perp_ticker_direction_funding_tooltip_short,
-                          })}
-                        </SizableText>
-                      </SizableText>
-                    ) : (
-                      <SizableText size="$bodySmMedium" color="$text">
-                        <SizableText size="$bodySmMedium" color="$red11">
-                          {intl.formatMessage({
-                            id: ETranslations.perp_ticker_direction_funding_tooltip_short,
-                          })}
-                        </SizableText>{' '}
-                        {intl.formatMessage({
-                          id: ETranslations.perp_ticker_direction_funding_tooltip_pays,
-                        })}{' '}
-                        <SizableText size="$bodySmMedium" color="$green11">
-                          {intl.formatMessage({
-                            id: ETranslations.perp_ticker_direction_funding_tooltip_long,
-                          })}
-                        </SizableText>
-                      </SizableText>
-                    )}
-                  </YStack>
-                </YStack>
-              }
-            />
-          </XStack>
-        </SkeletonContainer>
-      </YStack>
-    </DebugRenderTracker>
-  );
-}
+import { PerpTickerBarDesktop } from './PerpTickerBarDesktop';
+import { PerpTickerBarMobile } from './PerpTickerBarMobile';
 
 function PerpTickerBar() {
   const { gtMd } = useMedia();
-  const navigation = useAppNavigation();
-  const [asset] = usePerpsActiveAssetAtom();
-  const coin = asset?.coin || '';
-  const intl = useIntl();
-  const onPressCandleChart = useCallback(() => {
-    navigation.push(EModalPerpRoutes.MobilePerpMarket);
-  }, [navigation]);
-  const mmr = useMMR();
-  const [accountSummary] = usePerpsActiveAccountSummaryAtom();
-  const onPressTokenSelector = useCallback(() => {
-    navigation.pushModal(EModalRoutes.PerpModal, {
-      screen: EModalPerpRoutes.MobileTokenSelector,
-    });
-  }, [navigation]);
 
   if (!gtMd) {
-    const content = (
-      <XStack
-        flex={1}
-        bg="$bgApp"
-        gap="$4"
-        px="$4"
-        pt="$2"
-        pb="$1.5"
-        alignItems="flex-start"
-        justifyContent="space-between"
-      >
-        <YStack gap="$1">
-          <XStack
-            gap="$1"
-            bg="$bgApp"
-            onPress={onPressTokenSelector}
-            justifyContent="center"
-            alignItems="center"
-          >
-            <SizableText size="$headingXl">{coin}USD</SizableText>
-            <Badge radius="$1" bg="$bgSubdued" px="$1" py={0}>
-              <SizableText color="$textSubdued" fontSize={11}>
-                {intl.formatMessage({
-                  id: ETranslations.perp_label_perp,
-                })}
-              </SizableText>
-            </Badge>
-            <Icon name="ChevronTriangleDownSmallOutline" size="$5" />
-          </XStack>
-          <TickerBarChange24hPercent />
-        </YStack>
-
-        <XStack pt="$0.5" gap="$3" alignItems="center">
-          <Popover
-            title="MMR"
-            renderTrigger={
-              mmr ? (
-                <XStack
-                  borderRadius="$full"
-                  alignItems="center"
-                  gap="$1"
-                  px="$2"
-                  borderColor={
-                    parseFloat(mmr.multipliedBy(100).toFixed(2)) <= 50
-                      ? '$green11'
-                      : '$red11'
-                  }
-                  borderWidth="$px"
-                >
-                  <SizableText
-                    lineHeight={18}
-                    fontSize={10}
-                    color={
-                      parseFloat(mmr.multipliedBy(100).toFixed(2)) <= 50
-                        ? '$green11'
-                        : '$red11'
-                    }
-                  >
-                    {mmr.multipliedBy(100).toFixed(2)}%
-                  </SizableText>
-                </XStack>
-              ) : null
-            }
-            renderContent={
-              <YStack
-                bg="$bg"
-                justifyContent="center"
-                w="100%"
-                px="$5"
-                pt="$2"
-                pb="$5"
-                gap="$5"
-              >
-                <XStack alignItems="center" justifyContent="space-between">
-                  <YStack w="50%">
-                    <SizableText size="$bodyMd" color="$textSubdued">
-                      {intl.formatMessage({
-                        id: ETranslations.perp_account_panel_account_maintenance_margin,
-                      })}
-                    </SizableText>
-
-                    <PerpsAccountNumberValue
-                      value={accountSummary?.crossMaintenanceMarginUsed ?? ''}
-                      skeletonWidth={70}
-                      textSize="$bodyMdMedium"
-                    />
-                  </YStack>
-                  <YStack w="50%">
-                    <SizableText size="$bodyMd" color="$textSubdued">
-                      {intl.formatMessage({
-                        id: ETranslations.perp_account_cross_margin_ration,
-                      })}
-                    </SizableText>
-                    <SizableText
-                      size="$bodyMdMedium"
-                      color={
-                        parseFloat(mmr?.multipliedBy(100).toFixed(2) || '0') <=
-                        50
-                          ? '$green11'
-                          : '$red11'
-                      }
-                    >
-                      {mmr?.multipliedBy(100).toFixed(2)}%
-                    </SizableText>
-                  </YStack>
-                </XStack>
-
-                <Divider />
-                <YStack gap="$2">
-                  <SizableText size="$bodySmMedium">
-                    {intl.formatMessage({
-                      id: ETranslations.perp_account_panel_account_maintenance_margin_tooltip,
-                    })}
-                  </SizableText>
-                  <SizableText size="$bodySmMedium">
-                    {intl.formatMessage({
-                      id: ETranslations.perp_account_cross_margin_ration_tip,
-                    })}
-                  </SizableText>
-                </YStack>
-              </YStack>
-            }
-          />
-          <IconButton
-            icon="TradingViewCandlesOutline"
-            size="small"
-            iconProps={{ color: '$iconSubdued' }}
-            variant="tertiary"
-            onPress={onPressCandleChart}
-          />
-          <PerpSettingsButton testID="perp-mobile-settings-button" />
-        </XStack>
-      </XStack>
-    );
-    return (
-      <DebugRenderTracker name="PerpTickerBarMobile" position="top-right">
-        {content}
-      </DebugRenderTracker>
-    );
+    return <PerpTickerBarMobile />;
   }
-
-  const content = (
-    <XStack
-      bg="$bgApp"
-      borderBottomWidth="$px"
-      borderBottomColor="$borderSubdued"
-      p="$4"
-      alignItems="center"
-      justifyContent="flex-start"
-      gap="$6"
-      h={62}
-    >
-      <XStack gap="$4" alignItems="center">
-        <PerpTokenSelector />
-        <XStack alignItems="center" width={140} gap="$1.5" cursor="default">
-          <TickerBarMarkPrice />
-          <TickerBarChange24hPercent />
-        </XStack>
-      </XStack>
-
-      {/* Right: Market Data */}
-      <ScrollView
-        cursor="default"
-        horizontal
-        flex={1}
-        contentContainerStyle={{
-          gap: '$8',
-          alignItems: 'center',
-          justifyContent: 'flex-start',
-        }}
-      >
-        <TickerBarOraclePrice />
-        <TickerBar24hVolume />
-        <TickerBarOpenInterest />
-        <TickerBarFundingRate />
-      </ScrollView>
-    </XStack>
-  );
-  return (
-    <DebugRenderTracker name="PerpTickerBarDesktop" position="top-right">
-      {content}
-    </DebugRenderTracker>
-  );
+  return <PerpTickerBarDesktop />;
 }
 
 const PerpTickerBarMemo = memo(PerpTickerBar);
